Fixed issue 7: importing file with date and time column does not work
Try importing a file like:

"date";"time";"category";"duration";"distance";"hrm";"maxheartrate"
31.12.2010;16:00:00;"Laufen";00:45:29;8;127;0
30.12.2010;09:30:00;"Skitour/Bergsteigen";06:00:00;"18,5";130;0
29.12.2010;08:00:00;"Laufen";01:20:18;"13,5";129;0
29.12.2010;18:00:00;"Krafttraining";00:40:00;0;0;0
28.12.2010;17:00:00;"Laufen";00:59:44;12;145;158

It is imported as one single activity.

Problem: As a side effect from the bugfix for  issue 4 , the importer assumes that the time column contains a valid date - but it does not. As a result the importer assigns all lines to the same, wrong date.